requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.1
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
idna==3.10
PyJWT==2.9.0
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
gunicorn==23.0.0